결제 상세 UI, API 구현. 리뷰 이동/삭제 스와이핑 구현

+++ sososhopping_customer/app/src/main/java/com/sososhopping/customer/shop/view/ShopMainFragment.java
@@ -65,7 +65,7 @@ public class ShopMainFragment extends Fragment {
 
         //ViewModel Setting -> Main에서 유지되게
         shopInfoViewModel = new ViewModelProvider(this).get(ShopInfoViewModel.class);
-        shopInfoViewModel.setShopId(ShopMainFragmentArgs.fromBundle(getArguments()).getStoreId());
+        shopInfoViewModel.setShopId((int)ShopMainFragmentArgs.fromBundle(getArguments()).getStoreId());
 
         if(shopInfoViewModel.getDistance().getValue() == null){
             shopInfoViewModel.getDistance().setValue(ShopMainFragmentArgs.fromBundle(getArguments()).getDistance());
